% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{create_peak_list_from_mzml}
\alias{create_peak_list_from_mzml}
\title{Create peak list from mzML data}
\usage{
create_peak_list_from_mzml(mzml, scans)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{scans}{numeric vector containing scans to extract into peak list relevant, preferably produced by `get_peak_scans` function}
}
\value{
nested list of all data
}
\description{
The function extracts the relevant information and sorts it into nested lists for
use in the uncertainty functions
}
\section{note}{
 This function does not check to see if MS2 data is relevant to a specific peak, the `scans` parameter should be calculated through the `get_peak_scans` function
}

