% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{EIC}
\alias{EIC}
\title{Generate an extracted ion chromatogram object for a specific m/z value}
\usage{
EIC(mzml, mass, error, minerror = 0.001, MS1 = TRUE, charge = "ALL")
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{mass}{numeric m/z value for generating the extracted ion chromatogram}

\item{error}{numeric relative error (in ppm)}

\item{minerror}{numeric minimum absolute error for extracted ion}

\item{MS1}{logical to include only MS1 (TRUE) or all MS levels (FALSE)}

\item{charge}{character specify the ionization polarity to include ("positive", "negative", or "ALL")}
}
\value{
named list object containing time (`time)`) and extracted ion intensity (`intensity`)
}
\description{
Generate an extracted ion chromatogram object for a specific m/z value
}
