% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{extract_ms}
\alias{extract_ms}
\title{Generate mass spectrum of one or more scans}
\usage{
extract_ms(mzml, scans, mz.round = 4)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{scans}{integer vector of the one or more scan indices to combine into a single average spectrum}

\item{mz.round}{integer value number of decimals to round the m/z values}
}
\value{
data.frame containing mass spectrum `mz` and `int` values
}
\description{
Generate mass spectrum of one or more scans
}
