% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzMLtoR.R
\name{get_float}
\alias{get_float}
\title{Get encoding float size from single binary data array}
\usage{
get_float(binaryDataArrayList)
}
\arguments{
\item{binaryDataArrayList}{list from binary data array within a single mzml scan}
}
\value{
character string of `float32bit` or `float64bit`
}
\description{
Get encoding float size from single binary data array
}
