% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{get_peak_scans}
\alias{get_peak_scans}
\title{Spectral Uncertainty Functions ----------------------------------------------------------
Get scan numbers for a specific m/z value}
\usage{
get_peak_scans(mzml, mass, peak_start, peak_end, isowidth = 0.7)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{mass}{numeric m/z value of interest}

\item{peak_start}{numeric time (in minutes) of the start of the chromatographic peak}

\item{peak_end}{numeric time (in minutes) of the end of the chromatographic peak}

\item{isowidth}{numeric instrument isolation width (in Da) of the precursor ion, see below note.}
}
\value{
numeric vector containing the scan numbers for the MS1 and relevant MS2 scans for the specific precursor ion
}
\description{
Spectral Uncertainty Functions ----------------------------------------------------------
Get scan numbers for a specific m/z value
}
\section{note}{
 for DDA, isowidth is the Q1 isolation width, for SWATH/SONAR the isowidth is half the scan window width, and for DIA/AIF isowidth should be `NA`
}

