% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{getbaseion}
\alias{getbaseion}
\title{Get the base ion m/z value for a specific scan
Internal function}
\usage{
getbaseion(mzml, i)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{i}{integer scan number}
}
\description{
Get the base ion m/z value for a specific scan
Internal function
}
