% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{getBIC}
\alias{getBIC}
\title{Get the base ion intensity for a specific scan
Internal function}
\usage{
getBIC(mzml, i)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{i}{integer scan number}
}
\description{
Get the base ion intensity for a specific scan
Internal function
}
