% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{getcharge}
\alias{getcharge}
\title{Get the ionization polarity for a specific scan
Internal function}
\usage{
getcharge(mzml, i)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{i}{integer scan number}
}
\description{
Get the ionization polarity for a specific scan
Internal function
}
