% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{getEIC_from_peaklist}
\alias{getEIC_from_peaklist}
\title{Get Extracted Ion Chromatogram (EIC) from a peak list object
Internal function}
\usage{
getEIC_from_peaklist(
  peak,
  mass,
  masserror = 5,
  minerror = 0.002,
  mslevel = "MS1"
)
}
\arguments{
\item{peak}{list object result of the `create_peak_list` function}

\item{mass}{numeric m/z value of targeted mass}

\item{masserror}{numeric relative mass error (in ppm)}

\item{minerror}{numeric minimum absolute mass error (in Da)}

\item{mslevel}{character designating an EIC from MS1 ("MS1") or MS2 ("MS2") data}
}
\value{
data.frame object containing the EIC for a specified m/z value
}
\description{
Get Extracted Ion Chromatogram (EIC) from a peak list object
Internal function
}
