% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{getionint}
\alias{getionint}
\title{Base mzML analysis Functions ----------------------------------------------------------
Get the intensity of a specific m/z value for a specific scan
Internal function}
\usage{
getionint(mzml, i, minmass, maxmass)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{i}{integer scan number}

\item{minmass}{numeric minimum m/z value}

\item{maxmass}{numeric maximum m/z value}
}
\description{
Base mzML analysis Functions ----------------------------------------------------------
Get the intensity of a specific m/z value for a specific scan
Internal function
}
