% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{getmergedind}
\alias{getmergedind}
\title{Get indexes of m/z values to be merged together
Internal Function}
\usage{
getmergedind(masses, addms, masserror, minerror)
}
\arguments{
\item{masses}{numeric vector containing m/z values to be merged together}

\item{addms}{numeric data.frame containing m/z and intensity values to be merged into the masses and ints}

\item{masserror}{numeric relative mass error (in ppm)}

\item{minerror}{numeric minimum absolute mass error (in Da)}
}
\value{
integer vector containing m/z values to be merged together
}
\description{
Get indexes of m/z values to be merged together
Internal Function
}
