% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{getmslevel}
\alias{getmslevel}
\title{Get the MS level for a specific scan
Internal function}
\usage{
getmslevel(mzml, i)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{i}{integer scan number}
}
\description{
Get the MS level for a specific scan
Internal function
}
