% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{getprecursor}
\alias{getprecursor}
\title{Get the precursor ion m/z value for a specific scan
Internal function}
\usage{
getprecursor(mzml, i)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{i}{integer scan number}
}
\description{
Get the precursor ion m/z value for a specific scan
Internal function
}
