% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{getTIC}
\alias{getTIC}
\title{Get the total ion intensity for a specific scan
Internal function}
\usage{
getTIC(mzml, i)
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{i}{integer scan number}
}
\description{
Get the total ion intensity for a specific scan
Internal function
}
