% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{mergemass}
\alias{mergemass}
\title{Merges a set of m/z values of a mass spectra based on a list of m/z values
Internal function}
\usage{
mergemass(masses, ints, scans, addms, ind, masserror, minerror)
}
\arguments{
\item{masses}{numeric vector containing m/z values to be merged together}

\item{ints}{numeric vector containing int values to be merged together}

\item{scans}{integer vector}

\item{addms}{numeric data.frame containing m/z and intensity values to be merged into the masses and ints}

\item{ind}{integer index number of the scan}

\item{masserror}{numeric relative mass error (in ppm)}

\item{minerror}{numeric minimum absolute mass error (in Da)}
}
\value{
list object containing all merged masses
}
\description{
Merges a set of m/z values of a mass spectra based on a list of m/z values
Internal function
}
