% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{mergems}
\alias{mergems}
\title{Generate a merged mass spectrum comprised of multiple individual mass spectra
Internal function}
\usage{
mergems(mslist, peakindex, masserror = 5, minerror = 0.001)
}
\arguments{
\item{mslist}{list object containing multiple data.frames with individual mass spectra}

\item{peakindex}{integer representing the initial mass spectrum in the `mslist` to start merging process}

\item{masserror}{numeric relative mass error (in ppm)}

\item{minerror}{numeric minimum absolute mass error (in Da)}
}
\value{
list object containing merged masses, intensities, and scans
}
\description{
Generate a merged mass spectrum comprised of multiple individual mass spectra
Internal function
}
