% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzMLtoR.R
\name{mzMLtoR}
\alias{mzMLtoR}
\title{mzML I/O Functions ----------------------------------------------------------
Read mzML file into R environment}
\usage{
mzMLtoR(mzmlfile = file.choose())
}
\arguments{
\item{mzmlfile}{string path to .mzML file}
}
\value{
list object containing mzML data including metadata
}
\description{
Based function to read a standard mzML file into an R object. This includes unzipping any gzip-based spectra.
}
\section{note}{
 currently this format does not support the use of lockmass scans, this will be incorporated in future updates
}

\examples{
mzmlfile <- system.file('extdata', 'PFAS_STD.mzML', package = 'msuncertainty')
mzml <- mzMLtoR(mzmlfile)

}
