% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{peak_desc}
\alias{peak_desc}
\title{Get local minimum (peak start and end) in chromatographic peak
Internal Function}
\usage{
peak_desc(x, peak, width = 1, slope = 0.00044)
}
\arguments{
\item{x}{numeric vector representing a intensities of a chromatographic peak}

\item{peak}{integer index of local maximum}

\item{width}{integer number of points across peak to use to calculate slope}

\item{slope}{numeric slope limit that determines peak start and end}
}
\value{
numeric vector containing the index of the start and end of the peak
}
\description{
Get local minimum (peak start and end) in chromatographic peak
Internal Function
}
