% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{peak_width}
\alias{peak_width}
\title{Determine peak width at specific level of chromatographic peak
Internal Function}
\usage{
peak_width(chrom, peak, width = 100, slope = 1e-04, level = 0.1)
}
\arguments{
\item{chrom}{numeric vector representing a intensities of a chromatographic peak}

\item{peak}{integer index of local maximum}

\item{width}{integer number of points across peak to use to calculate slope}

\item{slope}{numeric slope limit that determines peak start and end}

\item{level}{numeric indicating the relative peak height to determine the width (0.1 means the width is determined at 10 percent of the peak height)}
}
\value{
numeric vector containing the index of the start and end of the peak at the specified peak height
}
\description{
Determine peak width at specific level of chromatographic peak
Internal Function
}
