% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{plot_ms}
\alias{plot_ms}
\title{Generate consensus mass spectrum plot}
\usage{
plot_ms(
  ms,
  xlim = NULL,
  ylim = NULL,
  main = "Mass Spectrum",
  color = "black",
  size = 1,
  removal = 0
)
}
\arguments{
\item{ms}{table uncertainty mass spectrum generated from `get_ums` function}

\item{xlim}{numeric vector containing the start and stop values of the x (m/z) axis (NULL if using default)}

\item{ylim}{numeric vector containing the start and stop values of the y (intensity) axis (NULL if using default)}

\item{main}{character string Title of the mass spectral plot}

\item{color}{character string color of the mass spectrum plot lines}

\item{size}{numeric width of the mass spectrum plot lines}

\item{removal}{numeric value of intensity to which remove mass spectrum peaks that are below specified intensity}
}
\value{
ggplot object
}
\description{
Generate consensus mass spectrum plot
}
