% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{slopeest}
\alias{slopeest}
\title{Calculate slope across points}
\usage{
slopeest(x, points)
}
\arguments{
\item{x}{numeric vector containing two or more values}

\item{points}{integer vector containing index of points of `x` to calculate slope}
}
\value{
numeric value of estimated slope.
}
\description{
Calculate slope across points
}
