% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ms.R
\name{TIC}
\alias{TIC}
\title{Generate a total ion chromatogram object}
\usage{
TIC(mzml, MS1 = TRUE, charge = "ALL")
}
\arguments{
\item{mzml}{list object parsed from mzML file using `mzMLtoR` function}

\item{MS1}{logical to include only MS1 (TRUE) or all MS levels (FALSE)}

\item{charge}{character specify the ionization polarity to include ("positive", "negative", or "ALL")}
}
\value{
named list object containing time (`time)`) and total ion intensity (`intensity`)
}
\description{
Generate a total ion chromatogram object
}
