% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzMLtoR.R
\name{unzipms}
\alias{unzipms}
\title{Unzip mzML mass spectra}
\usage{
unzipms(binary, compression = "none")
}
\arguments{
\item{binary}{character string representing}

\item{compression}{}
}
\value{
double vector containing unzipped MS data. If binary is empty, returns NULL.
}
\description{
Internal function used by mzMLtoR function to unzip mass spectral binary character strings.
}
