% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{write_ums}
\alias{write_ums}
\title{Write an uncertainty mass spectrum to a NIST .msp file}
\usage{
write_ums(
  ums,
  file,
  precursor = "",
  name = "R-exported Spectrum",
  include_uncertainty = TRUE
)
}
\arguments{
\item{ums}{uncertainty mass spectrum object generated by `get_ums` function}

\item{file}{character string to write the msp file to (include the extension)}

\item{precursor}{numeric value of the precursor ion m/z}

\item{name}{character string title to give the written mass spectrum}

\item{include_uncertainty}{logical to include the uncertainty calculations}
}
\value{
exports a mass spectrum to the designated file
}
\description{
Write an uncertainty mass spectrum to a NIST .msp file
}
