% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzMLtoR.R
\name{unzipms}
\alias{unzipms}
\title{Unzip mzML mass spectra}
\usage{
unzipms(binary, compression = "none", size = 8)
}
\arguments{
\item{binary}{character string representing}

\item{compression}{character listing the type of compression used ("none" or "gzip")}

\item{size}{integer float size (4 or 8)}
}
\value{
double vector containing unzipped MS data. If binary is empty, returns NULL.
}
\description{
Internal function used by mzMLtoR function to unzip mass spectral binary character strings.
}
